Compiles a math expression string into executable code and produces a delegate that represents the math expression.
