+++ MathEvaluation.Tests/Compilation/MathExpressionTests.CompileBoolean.cs
@@ -1,6 +1,5 @@
 ﻿using MathEvaluation.Context;
 using MathEvaluation.Extensions;
-using MathEvaluation.Parameters;
 
 namespace MathEvaluation.Tests.Compilation;
