+++ MathEvaluation.Tests/Compilation/MathExpressionTests_DecimalContext.Compile.cs
@@ -2,58 +2,36 @@
 using Xunit.Abstractions;
 using MathEvaluation.Context;
 using MathEvaluation.Extensions;
-using MathEvaluation.Parameters;
 
-namespace MathEvaluation.Tests;
+namespace MathEvaluation.Tests.Compilation;
 
-public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOutputHelper)
+public partial class MathExpressionTests_DecimalContext(ITestOutputHelper testOutputHelper)
 {
     private readonly DecimalScientificMathContext _scientificContext = new();
     private readonly DecimalProgrammingMathContext _programmingContext = new();
 
-    [Fact]
-    public void MathEvaluator_EvaluateDecimal_Null_ThrowArgumentNullException()
-    {
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(null!));
-        Assert.IsType<ArgumentNullException>(ex);
-    }
-
-    [Theory]
-    [InlineData("", "Expression string is empty or white space. (Parameter 'mathString')")]
-    [InlineData("   ", "Expression string is empty or white space. (Parameter 'mathString')")]
-    public void MathEvaluator_EvaluateDecimal_Empty_ThrowArgumentException(string expression,
-        string errorMessage)
-    {
-        testOutputHelper.WriteLine($"{expression}");
-
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression));
-        Assert.IsType<ArgumentException>(ex);
-        Assert.Equal(errorMessage, ex.Message);
-    }
-
     [Theory]
     [InlineData("+", "Error of evaluating the expression. It is not recognizable. Invalid token at position 1.")]
     [InlineData("-", "Error of evaluating the expression. It is not recognizable. Invalid token at position 1.")]
-    public void MathEvaluator_EvaluateDecimal_NotRecognizable_ThrowMathEvaluationException(string expression,
+    public void MathExpression_CompileDecimalThenInvoke_NotRecognizable_ThrowMathEvaluationException(string expression,
         string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression));
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).CompileDecimal());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
     [Fact]
-    public void MathEvaluator_EvaluateDecimal_DivideByZero_ThrowDivideByZeroException()
+    public void MathExpression_CompileDecimalThenInvoke_DivideByZero_ThrowDivideByZeroException()
     {
         testOutputHelper.WriteLine("0/0");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal("0/0"));
-        Assert.IsType<MathEvaluationException>(ex);
-        Assert.IsType<DivideByZeroException>(ex.InnerException);
+        var ex = Record.Exception(() => new MathExpression("0/0").CompileDecimal()());
+        Assert.IsType<DivideByZeroException>(ex);
 
-        Assert.Equal("Error of evaluating the expression. Attempted to divide by zero.", ex.Message);
+        Assert.Equal("Attempted to divide by zero.", ex.Message);
     }
 
     [Theory]
@@ -75,11 +53,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("2 - 5 * -10 / -2 / - 2 - 1", 2 - 5 * -10d / -2 / -2 - 1)]
     [InlineData("1 - -1", 1 - -1)]
     [InlineData("2 + \n(5 - 1) - \n\r 3", 2 + (5 - 1) - 3)]
-    public void MathEvaluator_EvaluateDecimal_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression);
+        var fn = new MathExpression(expression).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -89,11 +68,11 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("(3 * 2)//(2 * 2)", 1d)]
     [InlineData("6//-10", -1d)]
     [InlineData("4π // (3)π", 1d)]
-    public void MathEvaluator_EvaluateDecimal_HasFloorDivision_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasFloorDivision_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal(); var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -102,11 +81,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("(3 + 1)(5 + 2(5 - 1))", (3 + 1) * (5 + 2 * (5 - 1)))]
     [InlineData("(3 · 2)÷(5 × 2)", 3 * 2 / (5d * 2))]
     [InlineData("(3 + 1)(5 + 2(5 - 1))^2", (3 + 1) * (5 + 2 * (5 - 1)) * (5 + 2 * (5 - 1)))]
-    public void MathEvaluator_EvaluateDecimal_HasScientificNotation_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasScientificNotation_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -123,13 +103,14 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("22⹁888.32 ¤ * 30 / 323.34 / .5 - - 1 / (2 + 22⹁888.32 ¤) * 4 - 6", "ff-Adlm")]
     [InlineData("22 888,32 ¤ * 30 / 323,34 / ,5 - - 1 / (2 + 22 888,32 ¤) * 4 - 6", "fr")]
     [InlineData("22’888,32 ¤ * 30 / 323,34 / ,5 - - 1 / (2 + 22’888,32 ¤) * 4 - 6", "wae")]
-    public void MathEvaluator_EvaluateDecimal_HasNumbersInSpecificCulture_ExpectedValue(string expression, string? cultureName)
+    public void MathExpression_CompileDecimalThenInvoke_HasNumbersInSpecificCulture_ExpectedValue(string expression, string? cultureName)
     {
         var expectedValue = 22888.32m * 30 / 323.34m / .5m - -1 / (2 + 22888.32m) * 4 - 6;
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
         var cultureInfo = cultureName == null ? null : new CultureInfo(cultureName);
-        var value = MathEvaluator.EvaluateDecimal(expression, null, null, cultureInfo);
+        var fn = new MathExpression(expression, null, cultureInfo).CompileDecimal();
+        var value = fn();
 
         Assert.Equal(expectedValue, value);
     }
@@ -137,12 +118,13 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [Theory]
     [InlineData("|-2 CHF|^2 CHF * 30", "de-CH", 4d * 30)]
     [InlineData("|$-2|^$2 * 30", "en-US", 4d * 30)]
-    public void MathEvaluator_EvaluateDecimal_HasNumbersInSpecificCultureAsOperand_ExpectedValue(string expression, string? cultureName, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasNumbersInSpecificCultureAsOperand_ExpectedValue(string expression, string? cultureName, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
         var cultureInfo = cultureName == null ? null : new CultureInfo(cultureName);
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext, null, cultureInfo);
+        var fn = new MathExpression(expression, _scientificContext, cultureInfo).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -154,11 +136,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("-3**4", -81)]
     [InlineData("3 + 2(2 + 3.5)**2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
     [InlineData("3 + 2(2 + 3.5)  **2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathEvaluator_EvaluateDecimal_HasProgrammingPower_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasProgrammingPower_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _programmingContext);
+        var fn = new MathExpression(expression, _programmingContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -172,22 +155,24 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("-3^4sin(-PI/2)", 81d)]
     [InlineData("3 + 2(2 + 3.5)^ 2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
     [InlineData("3 + 2(2 + 3.5)  ^2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathEvaluator_EvaluateDecimal_HasScientificPower_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasScientificPower_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
 
     [Theory]
     [InlineData("4 % 3", 1)]
-    public void MathEvaluator_EvaluateDecimal_HasProgrammingModulus_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasProgrammingModulus_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _programmingContext);
+        var fn = new MathExpression(expression, _programmingContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -195,11 +180,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [Theory]
     [InlineData("4 mod 3", 1)]
     [InlineData("4 modulo 3 - 2", -1)]
-    public void MathEvaluator_EvaluateDecimal_HasScientificModulus_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasScientificModulus_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -212,11 +198,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("6.022e23", 6.022e23)] //Avogadro's number
     [InlineData("6.626e-34", 6.626e-34)] //Planck's constant
     [InlineData(".2E3 - 2", .2E3 - 2)]
-    public void MathEvaluator_EvaluateDecimal_HasExpNotationNumbers_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasExpNotationNumbers_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression);
+        var fn = new MathExpression(expression).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -225,11 +212,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("e", Math.E)]
     [InlineData("200e", 200 * Math.E)]
     [InlineData("200e^- 0.15", 172.14159528501156d)]
-    public void MathEvaluator_EvaluateDecimal_HasLnBase_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasLnBase_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -253,11 +241,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("abs(sin(-3))", 0.14112000805986721d)]
     [InlineData("|sin-3|", 0.14112000805986721d)]
     [InlineData("3 + 2|-2 + -3.5|  ^2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathEvaluator_EvaluateDecimal_HasAbs_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasAbs_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -275,11 +264,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("√9∛8", 6d)]
     [InlineData("∜16", 2d)]
     [InlineData("∜16∜16", 4d)]
-    public void MathEvaluator_EvaluateDecimal_HasRoot_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasRoot_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -294,11 +284,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("LNe", 1d)]
     [InlineData("ln100", 4.6051701859880918d)]
     [InlineData("-2ln(1/0.5 + √(1/0.5^2 + 1))", -2 * 1.4436354751788103d)]
-    public void MathEvaluator_EvaluateDecimal_HasLogarithmFn_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasLogarithmFn_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -320,11 +311,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("⌊sin3⌋", 0d)]
     [InlineData("⌊sin-3⌋", -1d)]
     [InlineData("3 + 2⌊2 + 3.5⌋  ^2", 3 + 2 * 25d)]
-    public void MathEvaluator_EvaluateDecimal_HasFloor_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasFloor_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -347,11 +339,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("⌈sin-3⌉", 0d)]
     [InlineData("⌈2 + 3.5⌉", 6d)]
     [InlineData("3 + 2⌈2 + 3.5⌉  ^2", 3 + 2 * 6d * 6d)]
-    public void MathEvaluator_EvaluateDecimal_HasCeiling_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasCeiling_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -367,32 +360,29 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("2!^(3)!^2!", 68719476736d)]
     [InlineData("2!^3^2!", 512d)]
     [InlineData("2!^(3)^2!", 512d)]
-    public void MathEvaluator_EvaluateDecimal_HasFactorial_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasFactorial_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var value = MathEvaluator.EvaluateDecimal(expression, _scientificContext);
+        var fn = new MathExpression(expression, _scientificContext).CompileDecimal();
+        var value = fn();
 
         Assert.Equal((decimal)expectedValue, value);
     }
 
     [Theory]
-    [InlineData("ln(1/x + √(1/x^2 + 1))", "x", 0.5, 1.4436354751788103d)]
-    [InlineData("x", "x", 0.5, 0.5d)]
-    [InlineData("2x", "x", 0.5, 1d)]
-    [InlineData("Math.PI", $"{nameof(Math)}.{nameof(Math.PI)}", Math.PI, Math.PI)]
-    public void MathEvaluator_EvaluateDecimal_HasVariable_ExpectedValue(string expression, string varName,
+    [InlineData("ln(1/x + √(1/x^2 + 1))", 0.5, 1.4436354751788103d)]
+    [InlineData("x", 0.5, 0.5d)]
+    [InlineData("2x", 0.5, 1d)]
+    [InlineData("PI", Math.PI, Math.PI)]
+    public void MathExpression_CompileDecimalThenInvoke_HasVariable_ExpectedValue(string expression,
         decimal varValue, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
-        testOutputHelper.WriteLine($"{varName} = {varValue}");
+        testOutputHelper.WriteLine($"variable value = {varValue}");
 
-        var parameters = new MathParameters();
-        parameters.BindVariable(varValue, varName);
-
-        var value = expression
-            .SetContext(_scientificContext)
-            .EvaluateDecimal(parameters);
+        var fn = expression.CompileDecimal(new { x = varValue, PI = varValue }, _scientificContext);
+        var value = fn(new { x = varValue, PI = varValue });
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -401,7 +391,7 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("getX1() + getX2( )", 0.5, 0.2, 0.5 + 0.2)]
     [InlineData("getX1()^2 + 2^getX2^2", 0.5, 3, 0.5 * 0.5 + 512)]
     [InlineData("ln(1/-getX1 + √(1/getX2^2 + 1))", -0.5, 0.5, 1.4436354751788103d)]
-    public void MathEvaluator_EvaluateDecimal_HasGetVariableFns_ExpectedValue(string expression,
+    public void MathExpression_CompileDecimalThenInvoke_HasGetVariableFns_ExpectedValue(string expression,
         decimal x1, decimal x2, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
@@ -410,11 +400,8 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
         var getX1 = () => x1;
         var getX2 = () => x2;
 
-        var parameters = new MathParameters(new { getX1, getX2 });
-
-        var value = expression
-            .SetContext(_scientificContext)
-            .EvaluateDecimal(parameters);
+        var fn = expression.CompileDecimal(new { getX1, getX2 }, _scientificContext);
+        var value = fn(new { getX1, getX2 });
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -422,7 +409,7 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [Theory]
     [InlineData("ln100 + x1 + x2", -1, 1, 4.6051701859880918d)]
     [InlineData("ln(1/-x1 + sqrt(1/(x2*x2) + 1))", -0.5, 0.5, 1.4436354751788103d)]
-    public void MathEvaluator_EvaluateDecimal_HasVariablesAndCustomLnSqrtFns_ExpectedValue(string expression,
+    public void MathExpression_CompileDecimalThenInvoke_HasVariablesAndCustomLnSqrtFns_ExpectedValue(string expression,
         decimal x1, decimal x2, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
@@ -431,11 +418,11 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
         var sqrt = Math.Sqrt;
         Func<double, double> ln = Math.Log;
 
-        var parameters = new MathParameters();
-        parameters.Bind(new { x1, x2, sqrt, ln });
+        var context = new MathContext(new { sqrt, ln });
+        var parameters = new { x1, x2 };
 
-        var value = expression
-            .EvaluateDecimal(parameters);
+        var fn = expression.CompileDecimal(parameters, context);
+        var value = fn(parameters);
 
         Assert.Equal((decimal)expectedValue, value);
     }
@@ -443,7 +430,7 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [Theory]
     [InlineData("1 + min(2, 1, 0.5, 4)", 1 + 0.5)]
     [InlineData("2min(-1, 1, 0.5, 4, 9999)", -2d)]
-    public void MathEvaluator_EvaluateDecimal_HasCustomMinFn_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_CompileDecimalThenInvoke_HasCustomMinFn_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
@@ -458,47 +445,45 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
             return minValue;
         };
 
-        var value = expression
-            .EvaluateDecimal(new { min });
+        var fn = expression.CompileDecimal(new { min });
+        var value = fn(new { min });
 
         Assert.Equal((decimal)expectedValue, value);
     }
 
     [Fact]
-    public void MathEvaluator_EvaluateDecimal_HasFactorialOfNotIntegerNumber_ThrowArgumentException()
+    public void MathExpression_CompileDecimalThenInvoke_HasFactorialOfNotIntegerNumber_ThrowArgumentException()
     {
         var expression = "0.2!";
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression, _scientificContext));
-        Assert.IsType<MathEvaluationException>(ex);
-        Assert.IsType<ArgumentException>(ex.InnerException);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).CompileDecimal()());
+        Assert.IsType<ArgumentException>(ex);
 
-        Assert.Equal("Error of evaluating the expression. Not integer number 0.2 isn't supported by the factorial function.", ex.Message);
-        Assert.Equal("Not integer number 0.2 isn't supported by the factorial function.", ex.InnerException.Message);
+        Assert.Equal("Not integer number 0.2 isn't supported by the factorial function.", ex.Message);
     }
 
     [Theory]
     [InlineData("1 + ctng(3 + 4)", "Error of evaluating the expression. 'ctng' is not recognizable. Invalid token at position 4.")]
     [InlineData("p", "Error of evaluating the expression. 'p' is not recognizable. Invalid token at position 0.")]
-    public void MathEvaluator_EvaluateDecimal_HasUnknowToken_ThrowMathEvaluationException(string expression,
+    public void MathExpression_CompileDecimalThenInvoke_HasUnknowToken_ThrowMathEvaluationException(string expression,
         string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression));
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).CompileDecimal());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
     [Fact]
-    public void MathEvaluator_EvaluateDecimal_HasIncorrectNumberFormat_ThrowFormatException()
+    public void MathExpression_CompileDecimalThenInvoke_HasIncorrectNumberFormat_ThrowFormatException()
     {
         var expression = "888e3.2";
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression));
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).CompileDecimal());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.IsType<FormatException>(ex.InnerException);
         Assert.Equal("Error of evaluating the expression. The input string '888e3.2' was not in a correct format.", ex.Message);
         Assert.Equal("The input string '888e3.2' was not in a correct format.", ex.InnerException.Message);
@@ -559,28 +544,26 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("Tanh^-1-1")]
     [InlineData("sinh(-∞)")]
     [InlineData("Sinh^-1(-∞)")]
-    public void MathEvaluator_EvaluateDecimal_ReturnsNanOrInfinity_ThrowOverflowException(string expression)
+    public void MathExpression_CompileDecimalThenInvoke_ReturnsNanOrInfinity_ThrowOverflowException(string expression)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression, _scientificContext));
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).CompileDecimal()());
 
-        Assert.IsType<MathEvaluationException>(ex);
-        Assert.IsType<OverflowException>(ex.InnerException);
-
-        Assert.Equal("Error of evaluating the expression. Value was either too large or too small for a Decimal.", ex.Message);
+        Assert.IsType<OverflowException>(ex);
+        Assert.Equal("Value was either too large or too small for a Decimal.", ex.Message);
     }
 
     [Theory]
     [InlineData("12 + 3 * (120 +5", "Error of evaluating the expression. It doesn't have the ')' closing symbol. Invalid token at position 9.")]
     [InlineData("abs(1/.5 + √(1/(0.5*0.5) + 1)", "Error of evaluating the expression. It doesn't have the ')' closing symbol. Invalid token at position 3.")]
     [InlineData("2 - 5 * ⌈-10 / 2 - 1", "Error of evaluating the expression. It doesn't have the '⌉' closing symbol. Invalid token at position 8.")]
-    public void MathEvaluator_EvaluateDecimal_ParenthesesOrFuncAreNotClosed_ThrowMathEvaluationException(string expression, string errorMessage)
+    public void MathExpression_CompileDecimalThenInvoke_ParenthesesOrFuncAreNotClosed_ThrowMathEvaluationException(string expression, string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression, _scientificContext));
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).CompileDecimal());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
@@ -588,12 +571,12 @@ public partial class MathEvaluatorTests_DecimalContext(ITestOutputHelper testOut
     [InlineData("12 + abs()", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 9.")]
     [InlineData("12 + abs / 1", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 8.")]
     [InlineData("0.5 + abs+1", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 9.")]
-    public void MathEvaluator_EvaluateDecimal_HasInvalidOperand_ThrowMathEvaluationException(string expression, string errorMessage)
+    public void MathExpression_CompileDecimalThenInvoke_HasInvalidOperand_ThrowMathEvaluationException(string expression, string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => MathEvaluator.EvaluateDecimal(expression, _scientificContext));
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).CompileDecimal());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 }
