+++ MathEvaluation.Tests/Evaluation/MathExpressionTests.Evaluate.cs
@@ -2,44 +2,25 @@
 using Xunit.Abstractions;
 using MathEvaluation.Context;
 using MathEvaluation.Extensions;
+using MathEvaluation.Parameters;
 
-namespace MathEvaluation.Tests.Compilation;
+namespace MathEvaluation.Tests.Evaluation;
 
 public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
 {
     private readonly ScientificMathContext _scientificContext = new();
     private readonly ProgrammingMathContext _programmingContext = new();
 
-    [Fact]
-    public void MathExpression_Null_ThrowArgumentNullException()
-    {
-        var ex = Record.Exception(() => new MathExpression(null!));
-        Assert.IsType<ArgumentNullException>(ex);
-    }
-
-    [Theory]
-    [InlineData("", "Expression string is empty or white space. (Parameter 'mathString')")]
-    [InlineData("   ", "Expression string is empty or white space. (Parameter 'mathString')")]
-    public void MathExpression_Empty_ThrowArgumentException(string expression,
-        string errorMessage)
-    {
-        testOutputHelper.WriteLine($"{expression}");
-
-        var ex = Record.Exception(() => new MathExpression(expression));
-        Assert.IsType<ArgumentException>(ex);
-        Assert.Equal(errorMessage, ex.Message);
-    }
-
     [Theory]
     [InlineData("+", "Error of evaluating the expression. It is not recognizable. Invalid token at position 1.")]
     [InlineData("-", "Error of evaluating the expression. It is not recognizable. Invalid token at position 1.")]
-    public void MathExpression_CompileThenInvoke_NotRecognizable_ThrowMathEvaluationException(string expression,
+    public void MathExpression_Evaluate_NotRecognizable_ThrowMathEvaluationException(string expression,
         string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression).Compile());
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
@@ -66,12 +47,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("2 - 5 * -10 / -2 / - 2 - 1", 2 - 5 * -10d / -2 / -2 - 1)]
     [InlineData("1 - -1", 1 - -1)]
     [InlineData("2 + \n(5 - 1) - \r\n 3", 2 + (5 - 1) - 3)]
-    public void MathExpression_CompileThenInvoke_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression).Compile();
-        var value = fn();
+        var value = new MathExpression(expression).Evaluate();
 
         testOutputHelper.WriteLine($"result: {value}");
 
@@ -83,12 +63,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("(3 * 2)//(2 * 2)", 1d)]
     [InlineData("6//-10", -1d)]
     [InlineData("4π // (3)π", 1d)]
-    public void MathExpression_CompileThenInvoke_HasFloorDivision_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasFloorDivision_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -97,12 +76,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("(3 + 1)(5 + 2(5 - 1))", (3 + 1) * (5 + 2 * (5 - 1)))]
     [InlineData("(3 · 2)÷(5 × 2)", 3 * 2 / (5d * 2))]
     [InlineData("(3 + 1)(5 + 2(5 - 1))^2", (3 + 1) * (5 + 2 * (5 - 1)) * (5 + 2 * (5 - 1)))]
-    public void MathExpression_CompileThenInvoke_HasScientificNotation_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasScientificNotation_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -119,14 +97,13 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("22⹁888.32 ¤ * 30 / 323.34 / .5 - - 1 / (2 + 22⹁888.32 ¤) * 4 - 6", "ff-Adlm")]
     [InlineData("22 888,32 ¤ * 30 / 323,34 / ,5 - - 1 / (2 + 22 888,32 ¤) * 4 - 6", "fr")]
     [InlineData("22’888,32 ¤ * 30 / 323,34 / ,5 - - 1 / (2 + 22’888,32 ¤) * 4 - 6", "wae")]
-    public void MathExpression_CompileThenInvoke_HasNumbersInSpecificCulture_ExpectedValue(string expression, string? cultureName)
+    public void MathExpression_Evaluate_HasNumbersInSpecificCulture_ExpectedValue(string expression, string? cultureName)
     {
         var expectedValue = 22888.32d * 30 / 323.34d / .5d - -1 / (2 + 22888.32d) * 4 - 6;
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
         var cultureInfo = cultureName == null ? null : new CultureInfo(cultureName);
-        var fn = new MathExpression(expression, null, cultureInfo).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, null, cultureInfo).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -137,15 +114,14 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("pow(,23, 2)", 0.23 * 0.23, "af")]
     [InlineData("pow( \r\n\t,23, 2)", 0.23 * 0.23, "af")]
     [InlineData("pow( \r\n\t,23, ,2 * 10)", 0.23 * 0.23, "af")]
-    public void MathExpression_CompileThenInvoke_HasCommaAsDecimalSeparatorInNumbers_ExpectedValue(string expression, double expectedValue, string cultureName)
+    public void MathExpression_Evaluate_HasCommaAsDecimalSeparatorInNumbers_ExpectedValue(string expression, double expectedValue, string cultureName)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
         var context = new MathContext();
         context.BindFunction(Math.Pow, "pow");
 
-        var fn = new MathExpression(expression, context, new CultureInfo(cultureName)).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, context, new CultureInfo(cultureName)).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -153,13 +129,12 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [Theory]
     [InlineData("|-2 CHF|^2 CHF * 30", "de-CH", 4d * 30)]
     [InlineData("|$-2|^$2 * 30", "en-US", 4d * 30)]
-    public void MathExpression_CompileThenInvoke_HasNumbersInSpecificCultureAsOperand_ExpectedValue(string expression, string? cultureName, double expectedValue)
+    public void MathExpression_Evaluate_HasNumbersInSpecificCultureAsOperand_ExpectedValue(string expression, string? cultureName, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
         var cultureInfo = cultureName == null ? null : new CultureInfo(cultureName);
-        var fn = new MathExpression(expression, _scientificContext, cultureInfo).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext, cultureInfo).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -173,12 +148,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("(-3)**0.5", double.NaN)]
     [InlineData("3 + 2(2 + 3.5)**2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
     [InlineData("3 + 2(2 + 3.5)  **2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathExpression_CompileThenInvoke_HasProgrammingPower_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasProgrammingPower_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _programmingContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _programmingContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -194,12 +168,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("(-3)^0.5", double.NaN)]
     [InlineData("3 + 2(2 + 3.5)^ 2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
     [InlineData("3 + 2(2 + 3.5)  ^2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathExpression_CompileThenInvoke_HasScientificPower_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasScientificPower_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -209,12 +182,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("3 - 4.5 % 3.1 / 3 * 2 + 4", 3 - 4.5 % 3.1 / 3 * 2 + 4)]
     [InlineData("3 - 4.5 % 3.1 * 3 * 2 + 4", 3 - 4.5 % 3.1 * 3 * 2 + 4)]
     [InlineData("3 - 2 / 4.5 % 3.1 / 3 * 2 + 4", 3 - 2 / 4.5 % 3.1 / 3 * 2 + 4)]
-    public void MathExpression_CompileThenInvoke_HasProgrammingModulus_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasProgrammingModulus_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _programmingContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _programmingContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -228,12 +200,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("4 modulo 3 - 2", -1)]
     [InlineData("3 - 4.5 Modulo 3.1 / 3 * 2 + 4", 3 - 4.5 % 3.1 / 3 * 2 + 4)]
     [InlineData("3 - 2 / 4.5 MOD 3.1 / 3 * 2 + 4", 3 - 2 / 4.5 % 3.1 / 3 * 2 + 4)]
-    public void MathExpression_CompileThenInvoke_HasScientificModulus_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasScientificModulus_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -248,12 +219,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData(".2E3 - 2", .2E3 - 2)]
     [InlineData("3 - 0.4e3 / 3 * 2 + 4", 3 - 0.4e3 / 3 * 2 + 4)]
     [InlineData(".1 - 0.4e3 / .3E-2 * .1E+10 + 2e+3", .1 - 0.4e3 / .3E-2 * .1E+10 + 2e+3)]
-    public void MathExpression_CompileThenInvoke_HasExpNotationNumbers_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasExpNotationNumbers_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression).Compile();
-        var value = fn();
+        var value = new MathExpression(expression).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -262,12 +232,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("e", Math.E)]
     [InlineData("200e", 200 * Math.E)]
     [InlineData("200e^- 0.15", 172.14159528501156d)]
-    public void MathExpression_CompileThenInvoke_HasLnBase_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasLnBase_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -292,12 +261,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("abs(sin(-3))", 0.14112000805986721d)]
     [InlineData("|sin-3|", 0.14112000805986721d)]
     [InlineData("3 + 2|-2 + -3.5|  ^2", 3 + 2 * (2 + 3.5d) * (2 + 3.5d))]
-    public void MathExpression_CompileThenInvoke_HasAbs_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasAbs_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -320,12 +288,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("∜-16", double.NaN)]
     [InlineData("∜16∜16", 4d)]
     [InlineData("1/√9^2", 1 / 9d)]
-    public void MathExpression_CompileThenInvoke_HasRoot_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasRoot_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -347,12 +314,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("ln-100", double.NaN)]
     [InlineData("ln(∞)", double.PositiveInfinity)]
     [InlineData("-2ln(1/0.5 + √(1/0.5^2 + 1))", -2 * 1.4436354751788103d)]
-    public void MathExpression_CompileThenInvoke_HasLogarithmFn_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasLogarithmFn_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -374,12 +340,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("⌊sin3⌋", 0d)]
     [InlineData("⌊sin-3⌋", -1d)]
     [InlineData("3 + 2⌊2 + 3.5⌋  ^2", 3 + 2 * 25d)]
-    public void MathExpression_CompileThenInvoke_HasFloor_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasFloor_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -402,12 +367,11 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("⌈sin-3⌉", 0d)]
     [InlineData("⌈2 + 3.5⌉", 6d)]
     [InlineData("3 + 2⌈2 + 3.5⌉  ^2", 3 + 2 * 6d * 6d)]
-    public void MathExpression_CompileThenInvoke_HasCeiling_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasCeiling_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
@@ -423,32 +387,31 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("2!^(3)!^2!", 68719476736d)]
     [InlineData("2!^3^2!", 512d)]
     [InlineData("2!^(3)^2!", 512d)]
-    public void MathExpression_CompileThenInvoke_HasFactorial_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasFactorial_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
-        var fn = new MathExpression(expression, _scientificContext).Compile();
-        var value = fn();
+        var value = new MathExpression(expression, _scientificContext).Evaluate();
 
         Assert.Equal(expectedValue, value);
     }
 
     [Theory]
-    [InlineData("ln(1/x + √(1/x^2 + 1))", 0.5, 1.4436354751788103d)]
-    [InlineData("x", 0.5, 0.5d)]
-    [InlineData("2x", 0.5, 1d)]
-    [InlineData("PI", Math.PI, Math.PI)]
-    [InlineData("2 * PI", Math.PI, 2 * Math.PI)]
-    public void MathExpression_CompileThenInvoke_HasVariable_ExpectedValue(string expression,
+    [InlineData("ln(1/x + √(1/x^2 + 1))", "x", 0.5, 1.4436354751788103d)]
+    [InlineData("x", "x", 0.5, 0.5d)]
+    [InlineData("2x", "x", 0.5, 1d)]
+    [InlineData("Math.PI", $"{nameof(Math)}.{nameof(Math.PI)}", Math.PI, Math.PI)]
+    [InlineData("2 * Math.PI", $"{nameof(Math)}.{nameof(Math.PI)}", Math.PI, 2 * Math.PI)]
+    public void MathExpression_Evaluate_HasVariable_ExpectedValue(string expression, string varName,
         double varValue, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
-        testOutputHelper.WriteLine($"variable value = {varValue}");
+        testOutputHelper.WriteLine($"{varName} = {varValue}");
 
-        var parameters = new { x = varValue, PI = varValue };
+        var parameters = new MathParameters();
+        parameters.BindVariable(varValue, varName);
 
-        var fn = expression.Compile(parameters, _scientificContext);
-        var value = fn(parameters);
+        var value = expression.Evaluate(parameters, _scientificContext);
 
         Assert.Equal(expectedValue, value);
     }
@@ -460,7 +423,7 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("sin2getX2()^2", 0.5, 3, 8.1836768414311347d)]
     [InlineData("SINgetX1^getX2", 0.5, 3, 0.12467473338522769d)]
     [InlineData("ln(1/-getX1 + √(1/getX2^2 + 1))", -0.5, 0.5, 1.4436354751788103d)]
-    public void MathExpression_CompileThenInvoke_HasGetValueFns_ExpectedValue(string expression,
+    public void MathExpression_Evaluate_HasGetValueFns_ExpectedValue(string expression,
         double x1, double x2, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
@@ -469,8 +432,9 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
         var getX1 = () => x1;
         var getX2 = () => x2;
 
-        var fn = expression.Compile(new { getX1, getX2 }, _scientificContext);
-        var value = fn(new { getX1, getX2 });
+        _scientificContext.Bind(new { getX1, getX2 });
+
+        var value = expression.Evaluate(null, _scientificContext);
 
         Assert.Equal(expectedValue, value);
     }
@@ -478,7 +442,7 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [Theory]
     [InlineData("ln100 + x1 + x2", -1, 1, 4.6051701859880918d)]
     [InlineData("ln(1/-x1 + sqrt(1/(x2*x2) + 1))", -0.5, 0.5, 1.4436354751788103d)]
-    public void MathExpression_CompileThenInvoke_HasVariablesAndCustomFns_ExpectedValue(string expression,
+    public void MathExpression_Evaluate_HasVariablesAndCustomFns_ExpectedValue(string expression,
         double x1, double x2, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
@@ -488,10 +452,7 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
         Func<double, double> ln = Math.Log;
 
         var context = new MathContext(new { sqrt, ln });
-        var parameters = new { x1, x2 };
-
-        var fn = expression.Compile(parameters, context);
-        var value = fn(parameters);
+        var value = expression.Evaluate(new { x1, x2 }, context); ;
 
         Assert.Equal(expectedValue, value);
     }
@@ -499,7 +460,7 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [Theory]
     [InlineData("1 + min(2, 1, 0.5, 4)", 1 + 0.5)]
     [InlineData("2min(-1, 1, 0.5, 4, 9999)", -2d)]
-    public void MathExpression_CompileThenInvoke_HasCustomMinFn_ExpectedValue(string expression, double expectedValue)
+    public void MathExpression_Evaluate_HasCustomMinFn_ExpectedValue(string expression, double expectedValue)
     {
         testOutputHelper.WriteLine($"{expression} = {expectedValue}");
 
@@ -514,44 +475,48 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
             return minValue;
         };
 
-        var fn = expression.Compile(new { min });
-        var value = fn(new { min });
+        var context = new MathContext();
+        context.Bind(new { min });
+
+        var value = expression.Evaluate(null, context);
 
         Assert.Equal(expectedValue, value);
     }
 
     [Fact]
-    public void MathExpression_CompileThenInvoke_HasFactorialOfNotIntegerNumber_ThrowArgumentException()
+    public void MathExpression_Evaluate_HasFactorialOfNotIntegerNumber_ThrowArgumentException()
     {
         var expression = "0.2!";
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Compile()());
-        Assert.IsType<ArgumentException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
+        Assert.IsType<ArgumentException>(ex.InnerException);
 
-        Assert.Equal("Not integer number 0.2 isn't supported by the factorial function.", ex.Message);
+        Assert.Equal("Error of evaluating the expression. Not integer number 0.2 isn't supported by the factorial function.", ex.Message);
+        Assert.Equal("Not integer number 0.2 isn't supported by the factorial function.", ex.InnerException.Message);
     }
 
     [Theory]
     [InlineData("1 + ctng(3 + 4)", "Error of evaluating the expression. 'ctng' is not recognizable. Invalid token at position 4.")]
     [InlineData("p", "Error of evaluating the expression. 'p' is not recognizable. Invalid token at position 0.")]
-    public void MathExpression_CompileThenInvoke_HasUnknowToken_ThrowMathEvaluationException(string expression, string errorMessage)
+    public void MathExpression_Evaluate_HasUnknowToken_ThrowMathEvaluationException(string expression, string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression).Compile());
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
     [Fact]
-    public void MathExpression_CompileThenInvoke_HasIncorrectNumberFormat_ThrowFormatException()
+    public void MathExpression_Evaluate_HasIncorrectNumberFormat_ThrowFormatException()
     {
         var expression = "888e3.2";
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression).Compile());
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.IsType<FormatException>(ex.InnerException);
         Assert.Equal("Error of evaluating the expression. The input string '888e3.2' was not in a correct format.", ex.Message);
         Assert.Equal("The input string '888e3.2' was not in a correct format.", ex.InnerException.Message);
@@ -561,12 +526,12 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("12 + 3 * (120 +5", "Error of evaluating the expression. It doesn't have the ')' closing symbol. Invalid token at position 9.")]
     [InlineData("abs(1/.5 + √(1/(0.5*0.5) + 1)", "Error of evaluating the expression. It doesn't have the ')' closing symbol. Invalid token at position 3.")]
     [InlineData("2 - 5 * ⌈-10 / 2 - 1", "Error of evaluating the expression. It doesn't have the '⌉' closing symbol. Invalid token at position 8.")]
-    public void MathExpression_CompileThenInvoke_ParenthesesOrFuncAreNotClosed_ThrowMathEvaluationException(string expression, string errorMessage)
+    public void MathExpression_Evaluate_ParenthesesOrFuncAreNotClosed_ThrowMathEvaluationException(string expression, string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Compile());
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 
@@ -574,12 +539,12 @@ public partial class MathExpressionTests(ITestOutputHelper testOutputHelper)
     [InlineData("12 + abs()", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 9.")]
     [InlineData("12 + abs / 1", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 8.")]
     [InlineData("0.5 + abs+1", "Error of evaluating the expression. The operand is not recognizable. Invalid token at position 9.")]
-    public void MathExpression_CompileThenInvoke_HasInvalidOperand_ThrowMathEvaluationException(string expression, string errorMessage)
+    public void MathExpression_Evaluate_HasInvalidOperand_ThrowMathEvaluationException(string expression, string errorMessage)
     {
         testOutputHelper.WriteLine($"{expression}");
 
-        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Compile());
-        Assert.IsType<MathEvaluationException>(ex);
+        var ex = Record.Exception(() => new MathExpression(expression, _scientificContext).Evaluate());
+        Assert.IsType<MathExpressionException>(ex);
         Assert.Equal(errorMessage, ex.Message);
     }
 }
