+++ MathEvaluation/CompilerServices/CallerArgumentExpressionAttribute.cs
@@ -1,4 +1,6 @@
-﻿namespace System.Runtime.CompilerServices
+﻿#pragma warning disable IDE0130 // Namespace does not match folder structure
+namespace System.Runtime.CompilerServices
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 {
     [AttributeUsage(AttributeTargets.Parameter)]
     internal sealed class CallerArgumentExpressionAttribute(string parameterName) : Attribute
