+++ MathEvaluation/MathEvaluation.csproj
@@ -7,22 +7,23 @@
         <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
         <Title>Math expression evaluator</Title>
         <Authors>Anton Antonov</Authors>
-        <Description>This .NET library offers a fast and comprehensive evaluator for dynamically evaluating mathematical expressions from strings. It supports a wide range of operations and allows for the use of custom variables, operators, and functions. The evaluator can be configured for different contexts, such as scientific or programming math expressions, making it highly versatile for various use cases. This flexibility, combined with its high performance, makes it an excellent choice for developers needing a robust mathematical evaluation tool.</Description>
+        <Description>The MathEvaluator .NET library allows you to evaluate and compile any mathematical expressions from a string dynamically. It supports a wide range of operations and allows for the use of custom variables, operators, and functions. The evaluator can be configured for different contexts, such as scientific or programming math expressions, making it highly versatile for various use cases. This flexibility, combined with its high performance, makes it an excellent choice for developers needing a robust mathematical evaluation tool.</Description>
         <GenerateDocumentationFile>True</GenerateDocumentationFile>
         <Copyright>Copyright © 2024 Anton Antonov</Copyright>
         <PackageProjectUrl>https://github.com/AntonovAnton/math.evaluation</PackageProjectUrl>
         <RepositoryUrl>https://github.com/AntonovAnton/math.evaluation</RepositoryUrl>
         <RepositoryType>git</RepositoryType>
         <PackageReadmeFile>README.md</PackageReadmeFile>
-        <PackageTags>math; math-expressions; evaluation; evaluator; mathematical-expressions; math-library; eval; expression-evaluator; math-parser; evaluate-expressions; math-expression-evaluator; math-expression-parser; mathexpression; math-eval-calc; csharp-maths; math-evaluator; math-evaluation;</PackageTags>
+        <PackageTags>math; math-expressions; evaluation; evaluator; mathematical-expressions; math-library; eval; ncalc; expression-evaluator; math-parser; evaluate-expressions; math-expression-evaluator; math-expression-parser; mathexpression; math-eval-calc; csharp-maths; math-evaluator; math-evaluation;</PackageTags>
         <PackageReleaseNotes>It targets .NET Standard 2.1 and higher version.
-Supports Boolean logic.
 Supports different mathematical contexts, such as scientific, programming, and other custom contexts.
 Evaluates to double, boolean, or decimal.
-Provides variable support within expressions.
-Extensible with custom functions.</PackageReleaseNotes>
+Compiles a math expression string into executable code and produces a delegate that represents the math expression.
+Provides variable support within math expressions.
+Extensible with custom functions and operators.
+        </PackageReleaseNotes>
         <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
-        <Version>1.1.10</Version>
+        <Version>2.0.0</Version>
         <PackageId>MathEvaluator</PackageId>
         <Product>MathEvaluator</Product>
     </PropertyGroup>
