+++ MathEvaluation/MathExpression.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Linq.Expressions;
 using MathEvaluation.Context;
 using MathEvaluation.Entities;
 using MathEvaluation.Extensions;
@@ -10,15 +9,13 @@ using MathEvaluation.Parameters;
 namespace MathEvaluation;
 
 /// <summary>
-/// Compiles any mathematical expression string to the <see cref="Func{T, TResult}"/> with parameters or to the <see cref="Func{TResult}"/>.
+/// Evaluates and compiles mathematical expressions from strings.
 /// </summary>
 public partial class MathExpression
 {
-    private static readonly Expression DoubleZero = Expression.Constant(0.0);
-
     private readonly NumberFormatInfo? _numberFormat;
+    private readonly char _decimalSeparator;
 
-    private ParameterExpression? _parameterExpression;
     private IMathParameters? _parameters;
 
     /// <summary>Gets the math expression string.</summary>
@@ -26,16 +23,12 @@ public partial class MathExpression
     public string MathString { get; }
 
     /// <summary>Gets the math context.</summary>
-    /// <value>The instance of the <see cref="IMathContext" /> interface.</value>
+    /// <value>The instance of the <see cref="IMathContext"/> interface.</value>
     public IMathContext? Context { get; }
 
     /// <summary>Gets the specified format provider.</summary>
     /// <value>The specified format provider.</value>
     public IFormatProvider? Provider { get; }
-
-    /// <summary>Gets the expression tree.</summary>
-    /// <value>The expression tree.</value>
-    public Expression? ExpressionTree { get; private set; }
 
     /// <summary>Initializes a new instance of the <see cref="MathExpression" /> class.</summary>
     /// <param name="mathString">The math expression string.</param>
@@ -54,47 +47,28 @@ public partial class MathExpression
         MathString = mathString;
         Context = context;
         Provider = provider;
+
         _numberFormat = provider != null ? NumberFormatInfo.GetInstance(provider) : null;
+        _decimalSeparator = _numberFormat?.NumberDecimalSeparator?.Length > 0 ? _numberFormat.NumberDecimalSeparator[0] : '.';
     }
 
-    /// <summary>Compiles the <see cref="MathString">math expression string</see>.</summary>
-    /// <returns>A delegate that represents the compiled expression.</returns>
-    /// <exception cref="MathEvaluationException"/>
-    public Func<double> Compile()
-    {
-        try
-        {
-            var i = 0;
-            ExpressionTree = Build(MathString, ref i, null, null, (int)EvalPrecedence.Unknown, false, DoubleZero);
-
-            return Expression.Lambda<Func<double>>(ExpressionTree).Compile();
-        }
-        catch (Exception ex)
-        {
-            throw CreateException(ex, MathString, Context, Provider, null);
-        }
-    }
-
-    /// <inheritdoc cref="Compile()"/>
-    /// <param name="parameters">The parameters of the <see cref="MathString">math expression string</see>.</param>
-    /// <exception cref="ArgumentNullException">parameters</exception>
+    /// <inheritdoc cref="Evaluate(IMathParameters?)"/>
     /// <exception cref="NotSupportedException">parameters</exception>
-    public Func<T, double> Compile<T>(T parameters)
+    public double Evaluate(object? parameters = null)
+        => Evaluate(parameters != null ? new MathParameters(parameters) : null);
+
+    /// <summary>Evaluates the <see cref="MathString">math expression string</see>.</summary>
+    /// <param name="parameters">The parameters of the <see cref="MathString">math expression string</see>.</param>
+    /// <returns>Value of the math expression.</returns>
+    /// <exception cref="MathExpressionException"/>
+    public double Evaluate(IMathParameters? parameters)
     {
-        if (parameters == null)
-            throw new ArgumentNullException(nameof(parameters));
+        _parameters = parameters;
 
         try
         {
-            _parameters = new MathParameters();
-            _parameters.Bind(parameters);
-
-            _parameterExpression = Expression.Parameter(typeof(T), nameof(parameters));
-
             var i = 0;
-            ExpressionTree = Build(MathString, ref i, null, null, (int)EvalPrecedence.Unknown, false, DoubleZero);
-
-            return Expression.Lambda<Func<T, double>>(ExpressionTree, _parameterExpression).Compile();
+            return Evaluate(MathString, ref i, null, null, (int)EvalPrecedence.Unknown);
         }
         catch (Exception ex)
         {
@@ -102,29 +76,27 @@ public partial class MathExpression
         }
     }
 
-    private Expression Build(ReadOnlySpan<char> mathString,
-        ref int i, char? separator, char? closingSymbol, int precedence, bool isOperand, Expression expression)
+    private double Evaluate(ReadOnlySpan<char> mathString,
+        ref int i, char? separator, char? closingSymbol, int precedence, bool isOperand = false, double value = default)
     {
-        var decimalSeparator = _numberFormat?.NumberDecimalSeparator?.Length > 0 ? _numberFormat.NumberDecimalSeparator[0] : '.';
-
         var start = i;
         while (mathString.Length > i)
         {
-            if (separator.HasValue && mathString.IsParamsSeparator(start, i, separator.Value, decimalSeparator) ||
+            if (separator.HasValue && mathString.IsParamsSeparator(start, i, separator.Value, _decimalSeparator) ||
                 closingSymbol.HasValue && mathString[i] == closingSymbol.Value)
             {
-                if (expression == DoubleZero)
+                if (value == default)
                     mathString.ThrowExceptionIfNotEvaluated(true, start, i);
 
-                return expression;
+                return value;
             }
 
-            if (mathString[i] is >= '0' and <= '9' || mathString[i] == decimalSeparator) //number
+            if (mathString[i] is >= '0' and <= '9' || mathString[i] == _decimalSeparator) //number
             {
                 if (isOperand)
-                    return Build(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Function, false, DoubleZero);
+                    return Evaluate(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Function);
 
-                expression = Expression.Constant(mathString.ParseNumber(_numberFormat, ref i));
+                value = mathString.ParseNumber(_numberFormat, ref i);
                 continue;
             }
 
@@ -132,125 +104,122 @@ public partial class MathExpression
             {
                 case '(':
                     if (precedence >= (int)EvalPrecedence.Function)
-                        return expression;
+                        return value;
 
                     var startParenthesis = i;
                     i++;
-                    var right = Build(mathString, ref i, null, ')', (int)EvalPrecedence.Unknown, false, DoubleZero);
+                    var result = Evaluate(mathString, ref i, null, ')', (int)EvalPrecedence.Unknown);
                     mathString.ThrowExceptionIfNotClosed(')', startParenthesis, ref i);
                     if (isOperand)
-                        return right;
+                        return result;
 
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, result);
+                    value = value == default ? result : value * result;
                     break;
                 case '+' when mathString.Length == i + 1 || mathString[i + 1] != '+':
                     if (isOperand || precedence >= (int)EvalPrecedence.LowestBasic && !mathString.IsMeaningless(start, i))
-                        return expression;
+                        return value;
 
                     i++;
-                    right = Build(mathString, ref i, separator, closingSymbol,
-                        precedence > (int)EvalPrecedence.LowestBasic ? precedence : (int)EvalPrecedence.LowestBasic, isOperand, DoubleZero);
-                    expression = Expression.Add(expression, right);
+                    var p = precedence > (int)EvalPrecedence.LowestBasic ? precedence : (int)EvalPrecedence.LowestBasic;
+                    value += Evaluate(mathString, ref i, separator, closingSymbol, p, isOperand);
                     if (isOperand)
-                        return expression;
+                        return value;
                     break;
                 case '-' when mathString.Length == i + 1 || mathString[i + 1] != '-':
-                    if (precedence >=(int)EvalPrecedence.LowestBasic && !mathString.IsMeaningless(start, i))
-                        return expression;
+                    if (precedence >= (int)EvalPrecedence.LowestBasic && !mathString.IsMeaningless(start, i))
+                        return value;
 
                     var isNegativity = start == i;
                     i++;
-                    right = Build(mathString, ref i, separator, closingSymbol,
-                        precedence > (int)EvalPrecedence.LowestBasic ? precedence : (int)EvalPrecedence.LowestBasic, isOperand, DoubleZero);
-                    expression = isNegativity ? Expression.Negate(right) : Expression.Subtract(expression, right); //it keeps sign
+                    p = precedence > (int)EvalPrecedence.LowestBasic ? precedence : (int)EvalPrecedence.LowestBasic;
+                    result = Evaluate(mathString, ref i, separator, closingSymbol, p, isOperand);
+                    value = isNegativity ? -result : value - result; //it keeps sign
                     if (isOperand)
-                        return expression;
+                        return value;
                     break;
                 case '*' when mathString.Length == i + 1 || mathString[i + 1] != '*':
                     if (precedence >= (int)EvalPrecedence.Basic)
-                        return expression;
+                        return value;
 
                     i++;
-                    right = Build(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic, false, DoubleZero);
-                    expression = Expression.Multiply(expression, right);
+                    value *= Evaluate(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic);
                     break;
                 case '/' when mathString.Length == i + 1 || mathString[i + 1] != '/':
                     if (precedence >= (int)EvalPrecedence.Basic)
-                        return expression;
+                        return value;
 
                     i++;
-                    right = Build(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic, false, DoubleZero);
-                    expression = Expression.Divide(expression, right);
+                    value /= Evaluate(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic);
                     break;
                 case ' ' or '\t' or '\n' or '\r': //space or tab or LF or CR
                     i++;
                     break;
                 default:
-                    var entity = Context?.FirstMathEntity(mathString[i..]) ?? _parameters?.FirstMathEntity(mathString[i..]);
-                    if (entity == null && _numberFormat != null && mathString.TryParseCurrency(_numberFormat, ref i))
+                    var entity = FirstMathEntity(mathString[i..]);
+                    if (entity == null && mathString.TryParseCurrency(_numberFormat, ref i))
                         break;
 
                     //highest precedence is evaluating first
                     if (precedence >= entity?.Precedence)
-                        return expression;
+                        return value;
 
-                    expression = BuildMathEntity(mathString, ref i, separator, closingSymbol, precedence, expression, entity);
+                    value = EvaluateMathEntity(mathString, ref i, separator, closingSymbol, precedence, value, entity);
                     if (isOperand)
-                        return expression;
+                        return value;
                     break;
             }
         }
 
-        if (expression == DoubleZero)
+        if (value == default)
             mathString.ThrowExceptionIfNotEvaluated(isOperand, start, i);
 
-        return expression;
+        return value;
     }
 
-    private Expression BuildOperand(ReadOnlySpan<char> mathString, ref int i, char? separator, char? closingSymbol)
+    private double EvaluateOperand(ReadOnlySpan<char> mathString, ref int i, char? separator, char? closingSymbol)
     {
         var start = i;
-        var expression = Build(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic, true, DoubleZero);
-        if (expression == DoubleZero)
+        var value = Evaluate(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic, true);
+        if (value == default)
             mathString.ThrowExceptionIfNotEvaluated(true, start, i);
 
-        return expression;
+        return value;
     }
 
-    private Expression BuildMathEntity(ReadOnlySpan<char> mathString,
-        ref int i, char? separator, char? closingSymbol, int precedence, Expression expression, IMathEntity? entity, bool throwError = true)
+    private double EvaluateMathEntity(ReadOnlySpan<char> mathString,
+        ref int i, char? separator, char? closingSymbol, int precedence, double value, IMathEntity? entity, bool throwError = true)
     {
         if (entity != null && entity.Precedence >= precedence)
         {
-            if (TryBuildEntity(mathString, entity, ref i, separator, closingSymbol, ref expression))
-                return expression;
+            if (TryEvaluateEntity(mathString, entity, ref i, separator, closingSymbol, ref value))
+                return value;
 
-            Expression convertExpression = expression == DoubleZero ? DecimalZero : Expression.Convert(expression, typeof(decimal));
-            if (TryBuildEntityDecimal(mathString, entity, ref i, separator, closingSymbol, ref convertExpression))
-                return Expression.Convert(convertExpression, typeof(double));
+            var decimalValue = (decimal)value;
+            if (TryEvaluateEntityDecimal(mathString, entity, ref i, separator, closingSymbol, ref decimalValue))
+                return (double)decimalValue;
         }
 
         if (throwError)
             mathString.ThrowExceptionInvalidToken(i);
 
-        return expression;
+        return value;
     }
 
-    private Expression BuildExponentiation(ReadOnlySpan<char> mathString,
-        ref int i, char? separator, char? closingSymbol, Expression expression)
+    private double EvaluateExponentiation(ReadOnlySpan<char> mathString,
+        ref int i, char? separator, char? closingSymbol, double value)
     {
         mathString.SkipMeaningless(ref i);
         if (mathString.Length <= i)
-            return expression;
+            return value;
 
         var precedence = (int)EvalPrecedence.Exponentiation;
-        var entity = Context?.FirstMathEntity(mathString[i..]) ?? _parameters?.FirstMathEntity(mathString[i..]);
-        return BuildMathEntity(mathString, ref i, separator, closingSymbol, precedence, expression, entity, false);
+        var entity = FirstMathEntity(mathString[i..]);
+        return EvaluateMathEntity(mathString, ref i, separator, closingSymbol, precedence, value, entity, false);
     }
 
-    private bool TryBuildEntity(ReadOnlySpan<char> mathString, IMathEntity entity,
-        ref int i, char? separator, char? closingSymbol, ref Expression expression)
+    private bool TryEvaluateEntity(ReadOnlySpan<char> mathString, IMathEntity entity,
+        ref int i, char? separator, char? closingSymbol, ref double value)
     {
         var start = i;
         switch (entity)
@@ -258,49 +227,42 @@ public partial class MathExpression
             case MathConstant<double> mathConstant:
                 {
                     i += entity.Key.Length;
-                    Expression right = Expression.Constant(mathConstant.Value);
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    var result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, mathConstant.Value);
+                    value = value == default ? result : value * result;
                     return true;
                 }
             case MathVariable<double> mathVariable:
                 {
                     i += entity.Key.Length;
-                    Expression right = Expression.Property(_parameterExpression, entity.Key);
-                    if (right.Type != typeof(double))
-                        right = Expression.Convert(right, typeof(double));
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    var result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, mathVariable.Value);
+                    value = value == default ? result : value * result;
                     return true;
                 }
             case MathOperandOperator<double> mathOperator:
                 {
                     i += entity.Key.Length;
-                    Expression<Func<double, double>> lambda = v => mathOperator.Fn(v);
-                    if (mathOperator.IsProcessingLeft)
-                        expression = Expression.Invoke(lambda, expression);
-                    else
-                    {
-                        var right = Build(mathString, ref i, separator, closingSymbol, (int)EvalPrecedence.Basic, false, DoubleZero);
-                        expression = Expression.Invoke(lambda, right);
-                    }
-                    expression = BuildExponentiation(mathString, ref i, separator, closingSymbol, expression);
+                    var fn = mathOperator.Fn;
+                    var result = mathOperator.IsProcessingLeft
+                        ? fn(value)
+                        : fn(EvaluateOperand(mathString, ref i, separator, closingSymbol));
+                    value = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, result);
                     return true;
                 }
             case MathOperandsOperator<double> mathOperator:
                 {
                     i += entity.Key.Length;
-                    Expression<Func<double, double, double>> lambda = (v1, v2) => mathOperator.Fn(v1, v2);
-                    var right = BuildOperand(mathString, ref i, separator, closingSymbol);
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = Expression.Invoke(lambda, expression, right);
+                    var fn = mathOperator.Fn;
+                    var right = EvaluateOperand(mathString, ref i, separator, closingSymbol);
+                    right = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, right);
+                    value = fn(value, right);
                     return true;
                 }
             case MathOperator<double> mathOperator:
                 {
                     i += entity.Key.Length;
-                    Expression<Func<double, double, double>> lambda = (v1, v2) => mathOperator.Fn(v1, v2);
-                    var right = Build(mathString, ref i, separator, closingSymbol, mathOperator.Precedence, false, DoubleZero);
-                    expression = Expression.Invoke(lambda, expression, right);
+                    var fn = mathOperator.Fn;
+                    var right = Evaluate(mathString, ref i, separator, closingSymbol, mathOperator.Precedence);
+                    value = fn(value, right);
                     return true;
                 }
             case MathGetValueFunction<double> mathFunction:
@@ -308,10 +270,9 @@ public partial class MathExpression
                     i += entity.Key.Length;
                     mathString.SkipParenthesis(ref i);
 
-                    Expression<Func<double>> lambda = () => mathFunction.Fn();
-                    Expression right = Expression.Invoke(lambda);
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    var result = mathFunction.Fn();
+                    result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, result);
+                    value = value == default ? result : value * result;
                     return true;
                 }
             case MathUnaryFunction<double> mathFunction:
@@ -321,16 +282,15 @@ public partial class MathExpression
                         mathString.ThrowExceptionIfNotOpened(mathFunction.OpeningSymbol.Value, start, ref i);
 
                     var arg = mathFunction.ClosingSymbol.HasValue
-                        ? Build(mathString, ref i, null, mathFunction.ClosingSymbol, (int)EvalPrecedence.Unknown, false, DoubleZero)
-                        : BuildOperand(mathString, ref i, separator, closingSymbol);
+                        ? Evaluate(mathString, ref i, null, mathFunction.ClosingSymbol, (int)EvalPrecedence.Unknown)
+                        : EvaluateOperand(mathString, ref i, separator, closingSymbol);
 
                     if (mathFunction.ClosingSymbol.HasValue)
                         mathString.ThrowExceptionIfNotClosed(mathFunction.ClosingSymbol.Value, start, ref i);
 
-                    Expression<Func<double, double>> lambda = (value) => mathFunction.Fn(value);
-                    Expression right = Expression.Invoke(lambda, arg);
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    var result = mathFunction.Fn(arg);
+                    result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, result);
+                    value = value == default ? result : value * result;
                     return true;
                 }
             case MathFunction<double> mathFunction:
@@ -338,10 +298,10 @@ public partial class MathExpression
                     i += entity.Key.Length;
                     mathString.ThrowExceptionIfNotOpened(mathFunction.OpeningSymbol, start, ref i);
 
-                    var args = new List<Expression>();
+                    var args = new List<double>();
                     while (mathString.Length > i)
                     {
-                        var arg = Build(mathString, ref i, mathFunction.Separator, mathFunction.ClosingSymbol, (int)EvalPrecedence.Unknown, false, DoubleZero);
+                        var arg = Evaluate(mathString, ref i, mathFunction.Separator, mathFunction.ClosingSymbol, (int)EvalPrecedence.Unknown);
                         args.Add(arg);
 
                         if (mathString[i] == mathFunction.Separator)
@@ -354,10 +314,9 @@ public partial class MathExpression
 
                     mathString.ThrowExceptionIfNotClosed(mathFunction.ClosingSymbol, start, ref i);
 
-                    Expression<Func<double[], double>> lambda = (values) => mathFunction.Fn(values);
-                    Expression right = Expression.Invoke(lambda, Expression.NewArrayInit(typeof(double), args));
-                    right = BuildExponentiation(mathString, ref i, separator, closingSymbol, right);
-                    expression = expression == DoubleZero ? right : Expression.Multiply(expression, right);
+                    var result = mathFunction.Fn([.. args]);
+                    result = EvaluateExponentiation(mathString, ref i, separator, closingSymbol, result);
+                    value = value == default ? result : value * result;
                     return true;
                 }
             default:
@@ -365,14 +324,17 @@ public partial class MathExpression
         }
     }
 
-    private static MathEvaluationException CreateException(Exception ex,
+    private static MathExpressionException CreateException(Exception ex,
         string mathString, IMathContext? context, IFormatProvider? provider, object? parameters)
     {
-        ex = ex is not MathEvaluationException ? new MathEvaluationException(ex.Message, ex) : ex;
+        ex = ex is not MathExpressionException ? new MathExpressionException(ex.Message, ex) : ex;
         ex.Data[nameof(mathString)] = mathString;
         ex.Data[nameof(context)] = context;
         ex.Data[nameof(provider)] = provider;
         ex.Data[nameof(parameters)] = parameters;
         throw ex;
     }
+
+    private IMathEntity? FirstMathEntity(ReadOnlySpan<char> mathString)
+        => Context?.FirstMathEntity(mathString) ?? _parameters?.FirstMathEntity(mathString);
 }
